+++ src/main/java/nextstep/oauth2/authentication/OAuth2AuthenticationToken.java
@@ -1,4 +1,4 @@
-package nextstep.oauth2;
+package nextstep.oauth2.authentication;
 
 import nextstep.oauth2.client.userinfo.OAuth2User;
 import nextstep.security.authentication.Authentication;
